Remove LineText.substring tracking
I assumed `str[a..b]` was an O(b) operation because it had to select
character number `b`, and some characters before `b` could be
multibyte, so Rust would have to inspect all of them which would take
O(b) time. In fact, `str[a..b]` selects *bytes* numbers `a..b`, which
is O(1) and means we can simplify the implementation here.